Re-organize helper functions
