+++ MarkEditMac/Modules/Sources/AppKitExtensions/Foundation/NSDocumentController+Extension.swift
@@ -1,0 +1,16 @@
+//
+//  NSDocumentController+Extension.swift
+//
+//  Created by cyan on 7/12/26.
+//
+
+import AppKit
+
+public extension NSDocumentController {
+  /**
+   Force the override of the last root directory for NSOpenPanel and NSSavePanel.
+   */
+  func setOpenPanelDirectory(_ directory: String) {
+    UserDefaults.standard.set(directory, forKey: "NSNavLastRootDirectory")
+  }
+}
