+++ MarkEditMac/Modules/Sources/AppKitExtensions/UI/NSPopover+Extension.swift
@@ -1,13 +1,12 @@
 //
 //  NSPopover+Extension.swift
-//  MarkEditMac
 //
 //  Created by cyan on 8/22/24.
 //
 
 import AppKit
 
-extension NSPopover {
+public extension NSPopover {
   var sourceView: NSView? {
     value(forKey: "positioningView") as? NSView
   }
