+++ MarkEditMac/Sources/Editor/Controllers/EditorViewController+Events.swift
@@ -6,6 +6,7 @@
 //
 
 import AppKit
+import AppKitExtensions
 import MarkEditKit
 
 extension EditorViewController {
