+++ MarkEditMac/Sources/Extras/NSDocumentController+Extras.swift
@@ -1,5 +1,5 @@
 //
-//  NSDocumentController+Extension.swift
+//  NSDocumentController+Extras.swift
 //  MarkEditMac
 //
 //  Created by cyan on 10/14/24.
@@ -25,13 +25,6 @@ extension NSDocumentController {
         }
       }
     }
-  }
-
-  /**
-   Force the override of the last root directory for NSOpenPanel and NSSavePanel.
-   */
-  func setOpenPanelDirectory(_ directory: String) {
-    UserDefaults.standard.set(directory, forKey: NSNavLastRootDirectory)
   }
 }
