+++ MarkEditMac/Sources/Extras/NSMenu+Extras.swift
@@ -1,5 +1,5 @@
 //
-//  NSMenu+Extension.swift
+//  NSMenu+Extras.swift
 //  MarkEditMac
 //
 //  Created by cyan on 5/26/25.
@@ -33,32 +33,6 @@ extension NSMenu {
         objc_AssociationPolicy.OBJC_ASSOCIATION_RETAIN_NONATOMIC
       )
     }
-  }
-
-  func firstActionNamed(_ title: String) -> NSMenuItem? {
-    firstDescendant {
-      $0.title == title && $0.action != nil
-    }
-  }
-
-  func firstMenuNamed(_ title: String) -> NSMenuItem? {
-    firstDescendant {
-      $0.title == title && $0.submenu != nil
-    }
-  }
-
-  func firstDescendant(matches predicate: (NSMenuItem) -> Bool) -> NSMenuItem? {
-    for item in items {
-      if predicate(item) {
-        return item
-      }
-
-      if let descendant = item.submenu?.firstDescendant(matches: predicate) {
-        return descendant
-      }
-    }
-
-    return nil
   }
 }
